move tiptapeditor back to own file
